orders/requirements.txt
Django==4.2.6
django-environ==0.11.2
django-filter==23.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2023.3
PyYAML==6.0.1
ujson==5.8.0
requests==2.31.0
